refactor: moved constants from globall file to locally scoped files
