+++ src/components/hamburger-menu/constants.ts
@@ -1,0 +1,26 @@
+import { Icons } from '../../shared/icons';
+export const menuIcons = [
+  {
+    linkTo: '/s-iteration-one',
+    icon: Icons.TelegramIcon,
+  },
+];
+
+export const menuTitlesArr = [
+  {
+    title: 'Парковка',
+    linkTo: '/s-iteration-one',
+  },
+  {
+    title: 'Страховка',
+    linkTo: '/s-iteration-one',
+  },
+  {
+    title: 'Бензин',
+    linkTo: '/s-iteration-one',
+  },
+  {
+    title: 'Обслуживание',
+    linkTo: '/s-iteration-one',
+  },
+];
