+++ src/pages/order-page/constants.ts
@@ -1,0 +1,18 @@
+export const breadcrumbsArr = [
+  {
+    title: 'Местоположение',
+    linkTo: '/simbirsoft-iteration-one/order',
+  },
+  {
+    title: 'Модель',
+    linkTo: '/simbirsoft-iteration-one/order',
+  },
+  {
+    title: 'Дополнительно',
+    linkTo: '/simbirsoft-iteration-one/order',
+  },
+  {
+    title: 'Итого',
+    linkTo: '/simbirsoft-iteration-one/order',
+  },
+];
