--- src/shared/constants/constants.ts
@@ -1,47 +1,0 @@
-import { Icons } from '../icons';
-
-export const menuIcons = [
-  {
-    linkTo: '/s-iteration-one',
-    icon: Icons.TelegramIcon,
-  },
-];
-
-export const menuTitlesArr = [
-  {
-    title: 'Парковка',
-    linkTo: '/s-iteration-one',
-  },
-  {
-    title: 'Страховка',
-    linkTo: '/s-iteration-one',
-  },
-  {
-    title: 'Бензин',
-    linkTo: '/s-iteration-one',
-  },
-  {
-    title: 'Обслуживание',
-    linkTo: '/s-iteration-one',
-  },
-];
-
-
-export const breadcrumbsArr = [
-    {
-      title: 'Местоположение',
-      linkTo: '/simbirsoft-iteration-one/order',
-    },
-    {
-      title: 'Модель',
-      linkTo: '/simbirsoft-iteration-one/order',
-    },
-    {
-      title: 'Дополнительно',
-      linkTo: '/simbirsoft-iteration-one/order',
-    },
-    {
-      title: 'Итого',
-      linkTo: '/simbirsoft-iteration-one/order',
-    },
-  ];
