--- src/shared/constants/index.ts
@@ -1,1 +1,0 @@
-export { menuTitlesArr, menuIcons, breadcrumbsArr } from './constants';
